refactors flight search to component
